requirements.txt
colorthief==0.2.1
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.0
joblib==1.4.2
kiwisolver==1.4.5
numpy==2.0.0
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
threadpoolctl==3.5.0
